requirements.txt
Bootstrap_Flask==2.2.0
requests=2.32.3
flask=3.1.0
python-dotenv~=1.0.1
Flask==2.3.2